+++ DdokBaro/Classes/Haptics.swift
