+++ DdokBaro/Classes/Vibrate.swift
