+++ DdokBaro/Model/Model.swift
@@ -10,3 +10,35 @@ import Foundation
 struct airPod{
     var degree:Int
 }
+
+struct DropDegree {
+    static let dropWhenBad: CGFloat = 5.0
+    static let dropWhenDanger: CGFloat = 10.0
+    static let dropWhenWorst: CGFloat = 15.0
+}
+
+
+//struct asd {
+//    private var currentWeight = (0.0, 0) // 현재 측정 각도
+//    private var userWeight = (0.0, 0) // 사용자 설정 가중치
+//    private var intPitch: Int = 0 // 목 기울기(정수)
+//    private var isZero = false
+//    private var nextView = false
+//    
+//    
+//    func checkNextView() -> Bool {
+//        let adsfadf
+//        return True`
+//    }
+//    
+//    func modifieruserWeight(sDF???) {
+//        
+//    }
+//    
+//    func asdasd() -> (Int, Int) {
+//        
+//        self.currentWeight = (10, 0)
+//        
+//        return currentWeight
+//    }
+//}
