--- DdokBaro/ViewControllers/HeadphoneMotion.swift
@@ -1,8 +1,0 @@
-//
-//  HeadphoneMotion.swift
-//  DdokBaro
-//
-//  Created by yusang on 2023/08/31.
-//
-
-import Foundation
