+++ DdokBaro/ViewControllers/MainView/MainViewController+event.swift
@@ -17,23 +17,30 @@ extension MainViewController {
     {
         let pitch = degrees(motion.attitude.pitch)
         intPitch = degreeInt(pitch)
-        //currentWeight = (pitch, degreeInt(pitch))
-        //print(intPitch)
+        
         DispatchQueue.main.async { [weak self] in
             if pitch > 0 {
                 self?.animationView2.setProgress(currentProgress: AnimationProgressTime(0))
             }
             else {
                 self?.animationView2.setProgress(currentProgress: AnimationProgressTime(max(-(pitch - userWeight.0)/40, 0)))
+                
+//                switch step {
+//                case .notgood:
+//                    
+//                case .bad:
+//                    
+//                default:
+//                    
+//                }
             }
-            
             //만약 목 각도가 정해진 기준 이상이면(notgood - 1단계, bad - 2단계, danger - 3단계)
             if intPitch - userWeight.1 < angle.notgood.rawValue {
-                
+
                 self?.animationView3.setPlay()
-                
+
                 if self!.isPause {
-                    self!.currentProgress -= self!.dropWhenBad * 0.00001
+                    self!.currentProgress -= DropDegree.dropWhenBad * 0.00001
                 }
                 if self!.currentProgress <= 0 {
                     isZero = true
@@ -41,28 +48,28 @@ extension MainViewController {
                     self!.currentProgress = 0
                     self!.createData()
                 }
-                
+
                 self!.waterWaveView.setupProgress(self!.currentProgress)
-                
+
                 if intPitch - userWeight.1 < angle.bad.rawValue {
                     self?.animationView3.setStop()
                     self?.animationView4.setPlay()
-                    
+
                     if !self!.isPause {
-                        self!.currentProgress -= self!.dropWhenWorst * 0.00001
+                        self!.currentProgress -= DropDegree.dropWhenWorst * 0.00001
                     }
-                    
+
                     self!.waterWaveView.setupProgress(self!.currentProgress)
-                    
+
                     if self!.motionTimer.isValid {
-                        
+
                     }
                     else {
                         self!.motionTimer = Timer.scheduledTimer(timeInterval: 180, target: self, selector: #selector(self?.badSound), userInfo: nil, repeats: false)
                     }
                 }
             }
-            
+
             //만약 목 각도가 기준선 이하로 돌아오면 타이머 삭제
             else{
                 self!.motionTimer.invalidate()
