+++ DdokBaro/ViewControllers/TestViewController.swift
@@ -7,6 +7,8 @@
 
 import UIKit
 
+
+//뭔지 모르겠음
 class TestViewController: UIViewController {
 
     override func viewDidLoad() {
